requirements.txt
gitpython
python-digitalocean
fuzzywuzzy